Validation, notifications, style update

+++ bank3/Bank/app/Http/Controllers/SaskaitaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Saskaita;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class SaskaitaController extends Controller
 {
@@ -42,7 +43,19 @@ class SaskaitaController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {
+    {   
+        $validator = Validator::make(
+            $request->all(),
+            [
+            'vardas' => 'required|alpha|min:3|max:20',
+            'pavarde' => 'required|alpha|min:3|max:20',
+            'ak' => 'required|min:11|max:11',
+            ]);
+
+            if ($validator->fails()) {
+                $request->flash();
+                return redirect()->back()->with('not', 'Neteisingai suvesti duomenys.');
+            }
         $sas = New Saskaita;
         $sas->vardas = $request->vardas;
         $sas->pavarde = $request->pavarde;
@@ -51,7 +64,7 @@ class SaskaitaController extends Controller
         $sas->s_nr = $request->s_nr;
         $sas->save();
 
-        return redirect()->route('bank-index');
+        return redirect()->route('bank-index')->with('ok','Nauja sąskaita sukurta.');
     }
 
     /**
@@ -87,6 +100,18 @@ class SaskaitaController extends Controller
      */
     public function update(Request $request, Saskaita $saskaita)
     {
+        $validator = Validator::make(
+            $request->all(),
+            [
+            'vardas' => 'required|alpha|min:3|max:20',
+            'pavarde' => 'required|alpha|min:3|max:20',
+            'ak' => 'required|min:11|max:11',
+            ]);
+
+            if ($validator->fails()) {
+                $request->flash();
+                return redirect()->back()->with('not', 'Neteisingai suvesti duomenys.');
+            }
         $saskaita->vardas = $request->vardas;
         $saskaita->pavarde = $request->pavarde;
         $saskaita->ak = $request->ak;
@@ -94,7 +119,7 @@ class SaskaitaController extends Controller
         $saskaita->s_nr = $request->s_nr;
         $saskaita->save();
 
-        return redirect()->route('bank-show', $saskaita);
+        return redirect()->route('bank-show', $saskaita)->with('ok', 'Sąskaitos informacija atnaujinta.');
     }
 
     /**
@@ -105,28 +130,35 @@ class SaskaitaController extends Controller
      */
     public function destroy(Saskaita $saskaita)
     {
+        if ($saskaita->suma > 0) {    
+            return redirect()->route('bank-show', $saskaita)->with('not', 'Sąskaitos su lėšomis ištrinti negalima');
+        }
         $saskaita->delete();
-        return redirect()->route('bank-index');
+        return redirect()->route('bank-index')->with('ok','Sąskaita '.$saskaita->s_nr.' ištrinta.');
     }
     public function plus(Request $request, Saskaita $saskaita){
+
+        if ($request->plus <= 0 || !is_numeric($request->plus)) {
+            return redirect()->route('bank-show', $saskaita)->with('not','Galima pridėti tik teigiamą sumą.');
+        }
         
         $saskaita->suma = $saskaita->suma + $request->plus;
         $saskaita->save();
 
-        return redirect()->route('bank-show', $saskaita);
+        return redirect()->route('bank-show', $saskaita)->with('ok','Prie sąskaitos pridėta: '.round($request->plus,2).' €');
     }
     public function minus(Request $request, Saskaita $saskaita){
-        if ($request->minus < 0) {
-            return redirect()->route('bank-show', $saskaita);
+        if ($request->minus < 0 || !is_numeric($request->minus)) {
+            return redirect()->route('bank-show', $saskaita)->with('not','Galima nurašyti tik teigiamą sumą.');
         }
         if ($saskaita->suma - $request->minus > 0) {
             $saskaita->suma = $saskaita->suma - $request->minus;
             $saskaita->save();
 
-            return redirect()->route('bank-show', $saskaita);
+            return redirect()->route('bank-show', $saskaita)->with('ok','Nuo sąskaitos nurašyta: '.round($request->minus, 2).' €');
         }
 
-        return redirect()->route('bank-show', $saskaita);
+        return redirect()->route('bank-show', $saskaita)->with('not','Nepakankamas likutis.');;
     }
     public function home(){
